task9_LocalStackProject/requirements.txt
pandas
boto3
s3fs
pyspark
apache-airflow-providers-amazon